# Cache header.sh and use loops to speedup unit tests