--- a/news/625-speedup-test
+++ b/news/625-speedup-test
@@ -1,0 +1,3 @@
+### Other
+
+* Speedup unit tests by factor 5. (#625)