--- a/news/625-speedup-test
+++ b/news/625-speedup-test
@@ -1,3 +1,19 @@
+### Enhancements
+
+* <news item>
+
+### Bug fixes
+
+* <news item>
+
+### Deprecations
+
+* <news item>
+
+### Docs
+
+* <news item>
+
 ### Other
 
 * Speedup unit tests by factor 5. (#625)